Implemented simple login just for testing, Userinfo has the same id as User, added some endpoints, swapped integers with longs for id

+++ backend-vfap/src/main/java/cz/osu/backendvfap/repository/EntryRepository.java
@@ -3,5 +3,5 @@ package cz.osu.backendvfap.repository;
 import cz.osu.backendvfap.model.Entry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntryRepository extends JpaRepository<Entry, Integer> {
+public interface EntryRepository extends JpaRepository<Entry, Long> {
 }
